[#13] Style: 유틸 함수 일부를 상수 파일로 이동
- grade 관련 객체들은 상수 파일에 위치하는게 맞는 거 같아서 위치 이동
- 유틸 파일에서 상수 파일 모듈 임포트 삭제
- dust 엘리먼트들을 객체로 만들어서 좀 더 의미를 가지게 변경
- forecast 파일 추가
