+++ FE/src/js/components/forecast.js
@@ -1,0 +1,4 @@
+import { _q, addClass, clearClass } from "../utils/utils";
+import { CLASS_NAME } from "../utils/constants";
+
+export const forecastContents = _q(`.${CLASS_NAME.forecastContents}`);
