+++ FE/src/js/utils/constants.js
@@ -52,3 +52,38 @@ export const API_URL = {
   dailyDustStatus: stationLocation => `${serverURL}${stationLocation}/daily-dust-status`,
   forecast: `${serverURL}forecast`,
 };
+
+export const getRestGradeClassName = className => {
+  const gradeClassList = [CLASS_NAME.grade1, CLASS_NAME.grade2, CLASS_NAME.grade3, CLASS_NAME.grade4];
+  return gradeClassList.filter(gradeClass => gradeClass !== className);
+};
+
+export const getGradeClassName = grade => {
+  const gradeClass = {
+    1: CLASS_NAME.grade1,
+    2: CLASS_NAME.grade2,
+    3: CLASS_NAME.grade3,
+    4: CLASS_NAME.grade4,
+  };
+  return gradeClass[grade] || null;
+};
+
+export const getGradeEmoji = grade => {
+  const gradeEmoji = {
+    1: "😀",
+    2: "🙂",
+    3: "😷",
+    4: "😱",
+  };
+  return gradeEmoji[grade] || null;
+};
+
+export const getGradeText = grade => {
+  const gradeText = {
+    1: "좋음",
+    2: "보통",
+    3: "나쁨",
+    4: "매우 나쁨",
+  };
+  return gradeText[grade] || null;
+};
